add WidgetHelper
